hugo: Watch for changes with `hugo watch`
In addition to changes made when Git repositories update, Hugo is able
to watch for timed changes to content (say, for content that has a
publication date in the future) and build these automatically.

This commit switches from a plain `hugo build` to a `hugo watch` process
to facilitate these use-cases.