dovecot: Only LIST directories in user maildir
This option has Dovecot only return directories in LIST commands, which
is a necessary workaround because of how our home and maildir locations
are the same, which sometimes has extraneous files (such as Sieve
scripts) appear in IMAP directory listings.

We should eventually move away from this unified location, but doing so
requires careful planning and migration.