postfix: Improve health-check command
This marginally improves the built-in health-check for Postfix by
ensuring that correct command pipelining is used, and that success is
predicated on command completion (rather than on the first response).

This, however, is still brittle, given the static `sleep` between
invocations; a preferable approach would have us ensure that
request/response pairs are matched.