rspamd: Improve handling of Redict timeouts
RSpamd retains DNS entries in its cache, which may or may not lead to
issues with resolve timeouts if Redict is independently restarted (e.g.
when updating versions).

This commit improves handling of timeouts by ensuring that Redict
resolution has a fairly low TTL of 1 second, while operations that
depend on Redict have a higher timeout; in this commit, we raise MX
check timeout to 5 seconds to compensate for any potential timeouts on
the Redict side.

In addition to the above, starting or restarting Redict will now attempt
to reload or restart RSpamd, which in turn should purge DNS cache, as a
workaround until a better solution is found upstream.