prosody: Deprecate `prosody-component-register` service
This commit moves component registration from a systemd service-based
solution to a mounted drop-in-based solution, similar to RSS2Email
feeds; this is due to ongoing issues with component configuration being
incorrectly processed on Prosody restart.

In order to facilitate more complex drop-in configurations, the
`container-environment` service will now process all `.env.template`
files in the specified service directory, which should help with
dropping in extra service configuration outside of the main directory.